Moved commands initialisation to its own file.
